Cover the scenario where the XVA was created in different region.
Sometime we need run fuel test in different network region. After
loading FM from XVA which was created in different network region,
it will have network access issue. This commit is resolving this
issue.

Also added some message in log to make the testing steps to be more
clear in the log file.